+++ testing/test_junitxml.py
@@ -610,14 +610,14 @@ def test_logxml_makedir(testdir):
 def test_escaped_parametrized_names_xml(testdir):
     testdir.makepyfile("""
         import pytest
-        @pytest.mark.parametrize('char', ["\\x00"])
+        @pytest.mark.parametrize('char', [u"\\x00"])
         def test_func(char):
             assert char
     """)
     result, dom = runandparse(testdir)
     assert result.ret == 0
     node = dom.find_first_by_tag("testcase")
-    node.assert_attr(name="test_func[#x00]")
+    node.assert_attr(name="test_func[\\x00]")
 
 
 def test_double_colon_split_function_issue469(testdir):